Add ViewModel/Room integration and refactor UI
Introduce ViewModel and Room-based data flow for gallery and quiz features: add GalleryViewModel, QuizViewModel, their factories, GalleryUiState and QuizUiState, and wire GalleryRepository into QuizApplication via an AppDatabase. Refactor GalleriActivity and QuizActivity to consume StateFlow UI state, handle loading/empty states, and call repository-backed actions (add/delete/submit/next). Fix GalleryRepository entity ID mapping, rename some dialog/state vars, and update UI components accordingly. Also update README task ownership, remove an IDE run-config selection, and bump AGP to 8.13.2.

+++ MyApplication/app/src/main/java/com/example/quizapp/data/GalleryRepository.kt
@@ -13,7 +13,7 @@ class GalleryRepository(
     // Leser fra Room og konverterer fra GalleryItemEntity → BildeOppforing
     private fun entityToBildeOppforing(entity: GalleryItemEntity): BildeOppforing {
         return BildeOppforing(
-            id = entity.id,
+            id = entity.id.toLong(),
             navn = entity.name,
             bildeUri = Uri.parse(entity.uri)
         )
